Cleaned up code and moved styling
