+++ client/src/components/answeredQuestions.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
-import { BlueText } from '../style.js';
+import { BlueText, Link } from '../style.js';
 
 const AnsweredQuestions = (props) => (
   <div>
-    <BlueText>
+    <Link href="CustomerQuestions">
+      <BlueText>
      | {props.numOfAnswers} answered questions
-    </BlueText>
+      </BlueText>
+    </Link>
   </div>
 );
